% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_todo_list.R
\name{create_todo_list}
\alias{create_todo_list}
\title{Create a To-Do List S3 Object}
\usage{
create_todo_list(todo_list = NULL, ...)
}
\arguments{
\item{todo_list}{A data frame containing the to-do list with columns: task, priority, and due.
If not provided, the function will import the current to-do list.}

\item{...}{additional arguments}
}
\value{
An instance of the TodoList S3 class.
}
\description{
This function creates an S3 object of class TodoList containing all entries
from the provided to-do list.
}
\examples{
# Example 1: Creating a TodoList object from an existing to-do list data frame:
list <- import_todo_list()
my_todo_list <- create_todo_list(list)

# Example 2: Creating a TodoList object from the current to-do list (imported from package data):
my_todo_list <- create_todo_list()

}
