% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_todo_list.R
\name{export_todo_list}
\alias{export_todo_list}
\title{Export To-Do List to CSV File}
\usage{
export_todo_list(todo_list, ...)
}
\arguments{
\item{todo_list}{The to-do list data frame or list to be exported.}

\item{...}{additional arguments}
}
\value{
Nothing. The to-do list is exported to a CSV file.
}
\description{
This function exports the updated to-do list back to a CSV file in the "inst" directory.
}
\examples{

list <- import_todo_list()

export_todo_list(list)

}
