% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.TodoList}
\title{Plot Todo-List}
\usage{
\method{plot}{TodoList}(x, ...)
}
\arguments{
\item{x}{An object of class 'TodoList'.}

\item{...}{additional arguments}
}
\value{
Nothing.
}
\description{
Plot Todo-List
}
\examples{

# Plotting TodoList objects:

list <- import_todo_list()
my_todo_list <- create_todo_list(list)
plot(my_todo_list)

}
