% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.TodoList}
\title{Print Todo-List}
\usage{
\method{print}{TodoList}(x, ...)
}
\arguments{
\item{x}{An object of class 'TodoList'}

\item{...}{additional arguments}
}
\value{
Nothing.
}
\description{
Print Todo-List
}
\examples{

# Printing TodoList objects:

list <- import_todo_list()
my_todo_list <- create_todo_list(list)
print(my_todo_list)

}
