% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_task.R
\name{remove_task}
\alias{remove_task}
\title{Remove Task Items from To-Do List}
\usage{
remove_task(task_index, todo_list, ...)
}
\arguments{
\item{task_index}{An integer specifying the index of the task to be removed.}

\item{todo_list}{A data frame or list representing the current to-do list.}

\item{...}{additional arguments}
}
\value{
An updated to-do list data frame or list with the specified item removed.
}
\description{
This function allows users to remove a specific item from their to-do list.
}
\examples{

list <- import_todo_list()
list <- add_task(list, "Buy bread", 3, "2024-03-15")

list <- remove_task(1, list)

}
