% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_items_by_date.R
\name{show_items_by_date.default}
\alias{show_items_by_date.default}
\title{Default method for showing items due on the specified date for other classes}
\usage{
\method{show_items_by_date}{default}(x, date = NULL, ...)
}
\arguments{
\item{x}{An object of other applicable classes.}

\item{date}{The date to filter by, in the format "YYYY-MM-DD".}

\item{...}{additional arguments}
}
\description{
Default method for showing items due on the specified date for other classes
}
