% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_items_by_date.R
\name{show_items_by_date}
\alias{show_items_by_date}
\title{Show Items Based on Date Filter Criteria}
\usage{
show_items_by_date(x, date = NULL, ...)
}
\arguments{
\item{x}{An object of class 'TodoList' or other applicable classes.}

\item{date}{The date to filter by, in the format "YYYY-MM-DD".}

\item{...}{additional arguments}
}
\value{
Nothing. The function displays items from a to-do list or other applicable classes based on the specified date filter criteria.
}
\description{
This generic function displays items from a to-do list or other applicable classes based on a specified date filter criteria.
}
\examples{

# Showing items due on the specified date for TodoList objects:

list <- import_todo_list()
my_todo_list <- create_todo_list(list)
show_items_by_date(my_todo_list, date = "2024-03-15")

# Showing items due on the specified date for other classes:

other_object <- list(task = paste("Task", 1:15),
priority = sample(1:5, 15, replace = TRUE),
due = sample(seq(as.Date("2022-01-01"), as.Date("2025-12-31"), by = "day"), 15))

show_items_by_date(other_object, date = "2022-01-01")

}
