% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_items_by_date.R
\name{show_items_by_date.TodoList}
\alias{show_items_by_date.TodoList}
\title{Method for showing items due on the specified date for TodoList objects}
\usage{
\method{show_items_by_date}{TodoList}(x, date = NULL, ...)
}
\arguments{
\item{x}{An object of class 'TodoList'.}

\item{date}{The date to filter by, in the format "YYYY-MM-DD".}

\item{...}{additional arguments}
}
\description{
Method for showing items due on the specified date for TodoList objects
}
