% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_items_by_priority.R
\name{show_items_by_priority.default}
\alias{show_items_by_priority.default}
\title{Default method for showing items with the specified priority for other classes}
\usage{
\method{show_items_by_priority}{default}(x, priority, ...)
}
\arguments{
\item{x}{An object of other applicable classes.}

\item{priority}{The priority level to filter by, an integer between 1 and 5.}

\item{...}{additional arguments}
}
\description{
Default method for showing items with the specified priority for other classes
}
