% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_items_by_priority.R
\name{show_items_by_priority}
\alias{show_items_by_priority}
\title{Show Items Based on Priority Filter Criteria}
\usage{
show_items_by_priority(x, priority, ...)
}
\arguments{
\item{x}{An object of class 'TodoList' or other applicable classes.}

\item{priority}{The priority level of the tasks to filter by, an integer between 1 and 5.}

\item{...}{additional arguments}
}
\value{
Nothing. The function displays items from a to-do list or other applicable classes based on the specified priority filter criteria.
}
\description{
This generic function displays items from a to-do list or other applicable classes based on a specified priority filter criteria.
}
\examples{

# Showing items with the specified priority for TodoList objects:
list <- import_todo_list()
my_todo_list <- create_todo_list(list)
show_items_by_priority(my_todo_list, priority = 3)

# Showing items with the specified priority for other classes:
other_object <- list(task = paste("Task", 1:15),
priority = sample(1:5, 15, replace = TRUE),
due = sample(seq(as.Date("2022-01-01"), as.Date("2025-12-31"), by = "day"), 15))

show_items_by_priority(other_object, priority = 2)


}
