% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_items_by_priority.R
\name{show_items_by_priority.TodoList}
\alias{show_items_by_priority.TodoList}
\title{Method for showing items with the specified priority for TodoList objects}
\usage{
\method{show_items_by_priority}{TodoList}(x, priority, ...)
}
\arguments{
\item{x}{An object of class 'TodoList'.}

\item{priority}{The priority level to filter by, an integer between 1 and 5.}

\item{...}{additional arguments}
}
\description{
Method for showing items with the specified priority for TodoList objects
}
