% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_today.R
\name{show_today.default}
\alias{show_today.default}
\title{Default method for showing items due today or earlier for other classes}
\usage{
\method{show_today}{default}(x, ...)
}
\arguments{
\item{x}{An object of other applicable classes.}

\item{...}{additional arguments}
}
\description{
Default method for showing items due today or earlier for other classes
}
