% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_today.R
\name{show_today}
\alias{show_today}
\title{Show Items due Today or Earlier}
\usage{
show_today(x, ...)
}
\arguments{
\item{x}{An object of class 'TodoList' or other applicable classes representing a to-do list.}

\item{...}{additional arguments}
}
\value{
Nothing. The function displays items due today or earlier from a to-do list,
accompanied by a bar plot illustrating task priorities.
}
\description{
This generic function shows items due today or earlier from a to-do list.
It also generates a bar plot showing the priority of tasks, with higher priority tasks indicated by taller bars.
}
\examples{

# Showing items due today or earlier for TodoList objects:

list <- import_todo_list()
my_todo_list <- create_todo_list(list)
show_today(my_todo_list)

# Showing items due today or earlier for objects of other classes:

other_object <- list(task = paste("Task", 1:15),
priority = sample(1:5, 15, replace = TRUE),
due = sample(seq(as.Date("2022-01-01"), as.Date("2025-12-31"), by = "day"), 15))

show_today(other_object)

}
