% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_today.R
\name{show_today.TodoList}
\alias{show_today.TodoList}
\title{Method for showing items due today or earlier for TodoList objects}
\usage{
\method{show_today}{TodoList}(x, ...)
}
\arguments{
\item{x}{An object of class 'TodoList'.}

\item{...}{additional arguments}
}
\description{
Method for showing items due today or earlier for TodoList objects
}
