% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_items.R
\name{sort_items.default}
\alias{sort_items.default}
\title{Default method for sorting items in other applicable classes}
\usage{
\method{sort_items}{default}(x, date = TRUE, ascending = TRUE, ...)
}
\arguments{
\item{x}{An object of other applicable classes.}

\item{date}{Logical, indicating whether to sort by date.}

\item{ascending}{Logical, indicating the sorting order.}

\item{...}{additional arguments}
}
\value{
The sorted object.
}
\description{
Default method for sorting items in other applicable classes
}
