% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_items.R
\name{sort_items.TodoList}
\alias{sort_items.TodoList}
\title{Method for sorting items in a TodoList object}
\usage{
\method{sort_items}{TodoList}(x, date = TRUE, ascending = TRUE, ...)
}
\arguments{
\item{x}{An object of class 'TodoList'.}

\item{date}{Logical, indicating whether to sort by date.}

\item{ascending}{Logical, indicating the sorting order.}

\item{...}{additional arguments}
}
\value{
The sorted TodoList object.
}
\description{
Method for sorting items in a TodoList object
}
