% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_task.R
\name{update_task}
\alias{update_task}
\title{Update Task Item in To-Do List}
\usage{
update_task(
  task_index,
  todo_list,
  new_task = NULL,
  new_priority = NULL,
  new_due = NULL,
  ...
)
}
\arguments{
\item{task_index}{An integer specifying the index of the task to be updated.}

\item{todo_list}{A data frame or list representing the current to-do list.}

\item{new_task}{A character string specifying the updated task description.}

\item{new_priority}{An integer specifying the updated priority of the task. Must be within the range 1-5.}

\item{new_due}{A Date object specifying the updated due date of the task.}

\item{...}{additional arguments}
}
\value{
An updated to-do list data frame or list with the specified item updated.
}
\description{
This function allows users to update the task description, priority, or due date of an existing item in their to-do list.
}
\examples{

list <- import_todo_list()
list <- add_task(list, "Buy bread", 3, "2024-03-15")

list <- update_task(1, list, "Buy milk", 4, "2024-03-18")

}
